pyproject.toml
[project]
name = "headlessperfbench"
version = "0.1.0"
description = "Performance and rendering fidelity benchmarks for Chrome headful, headless, and headless-shell modes"
readme = "README.md"
license = "AGPL-3.0-or-later"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "playwright>=1.58.0",
    "psutil>=7.2.2",
    "rich>=14.3.2",
    "scipy>=1.17.0",
    "uvicorn>=0.41.0",
]

[project.scripts]
hpb-analyze = "analysis.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["collector", "analysis"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
